// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  MODERATOR: 'MODERATOR',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const RoomStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  ARCHIVED: 'ARCHIVED'
} as const

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus]


export const DevicePlatform = {
  IOS: 'IOS',
  ANDROID: 'ANDROID',
  WEB: 'WEB'
} as const

export type DevicePlatform = (typeof DevicePlatform)[keyof typeof DevicePlatform]


export const MessageType = {
  TEXT: 'TEXT',
  IMAGE: 'IMAGE',
  VOICE_NOTE: 'VOICE_NOTE',
  VIDEO: 'VIDEO'
} as const

export type MessageType = (typeof MessageType)[keyof typeof MessageType]


export const MessageStatus = {
  SENT: 'SENT',
  DELIVERED: 'DELIVERED',
  READ: 'READ'
} as const

export type MessageStatus = (typeof MessageStatus)[keyof typeof MessageStatus]